item 1, 2, 4, 6, 8, 9, 11, 12, 13 have been fixed.
item 5,6 : the idea is that only one copy of the fact can be provided by each insurance company.
i.e. Company A cannot update fact as Company B.
item 10: That should be a separate workflow between the customer and his/her insurer as FNOL is accessible by all the parties.